TST: Moved derivative tests in separate file
