+++ test/test_topography1.py
@@ -1117,49 +1117,7 @@ def test_autocompletion(self):
     self.assertIn('to_uniform', dir(t3))
 
 
-def test_uniform_vs_nonuniform():
-    t1 = fourier_synthesis([12], [6], 0.8, rms_slope=0.1)
-    t2 = t1.to_nonuniform()
 
-    d1 = t1.derivative(1)
-    d2 = t2.derivative(1)
-
-    print(t1.nb_grid_pts)
-    print(d1.shape)
-    print(d2.shape)
-
-    np.testing.assert_allclose(d1[:-1], d2)
-
-
-def test_analytic():
-    nb_pts = 1488
-    s = 4 * np.pi
-    x = np.arange(nb_pts) * s / nb_pts
-    h = np.sin(x)
-    t1 = UniformLineScan(h, (s,))
-    t2 = UniformLineScan(h, (s,), periodic=True)
-    t3 = t1.to_nonuniform()
-
-    d1 = t1.derivative(1)
-    d2 = t2.derivative(1)
-    d3 = t3.derivative(1)
-
-    np.testing.assert_allclose(d1, np.cos(x[:-1] + (x[1] - x[0]) / 2),
-                               atol=1e-5)
-    np.testing.assert_allclose(d2, np.cos(x + (x[1] - x[0]) / 2),
-                               atol=1e-5)
-    np.testing.assert_allclose(d3, np.cos(x[:-1] + (x[1] - x[0]) / 2),
-                               atol=1e-5)
-
-    d1 = t1.derivative(2)
-    d2 = t2.derivative(2)
-    d3 = t3.derivative(2)
-
-    np.testing.assert_allclose(d1, -np.sin(x[:-2] + (x[1] - x[0])),
-                               atol=1e-5)
-    np.testing.assert_allclose(d2, -np.sin(x + (x[1] - x[0])), atol=1e-5)
-    np.testing.assert_allclose(d3, -np.sin(x[:-2] + (x[1] - x[0])),
-                               atol=1e-5)
 
 
 @pytest.mark.parametrize("ny", [7, 6])
